Update Length.java
Made length() function static.
Length () function needs to be static in order to be called from the 'main' function. In your code 'main' is a static function and length() is a non-static function. So the 'main' will not be able to call length(), as one static function cannot call another non-static function. Thus 'static' keyword needs to be added in front of 'length()' so 'main' can call it.